+++ src/agentlang/cli/constant.clj
@@ -8,7 +8,7 @@
 (def ^:const model-filename "model.al")
 (def ^:const al-file-extension ".al")
 (def ^:const git-deps-directory "deps/git")
-(def ^:const baseline-version "0.6.2")
+(def ^:const baseline-version "0.7.0-alpha1")
 
 (def ^:const env-var-github-username "GITHUB_USERNAME")
 (def ^:const env-var-github-token "GITHUB_TOKEN")